消除告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@16557 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/client/client.c
@@ -4722,7 +4722,7 @@ SDB_EXPORT INT32 sdbQuery1 ( sdbCollectionHandle cHandle,
 
    // query with return data
    if ( ((MsgHeader*)cs->_pReceiveBuffer)->messageLength >
-        ossRoundUpToMultipleX( sizeof(MsgOpReply), 4 ) )
+        (INT32)ossRoundUpToMultipleX( sizeof(MsgOpReply), 4 ) )
    {
       cursor->_pReceiveBuffer = cs->_pReceiveBuffer ;
       cs->_pReceiveBuffer = NULL ;
